requirements.txt
pillow~=10.3.0
numpy~=1.24.4
streamlit
langchain
google-cloud-vision
pytesseract
opencv-python
pyttsx3
